move Spectrum componentent into scenes/game
