+++ src/scenes/game/Spectrum.tsx
